requirements.txt
Flask==3.0.3
python-jose[cryptography]==3.3.0
cryptography==42.0.5
pytest==8.3.2
pytest-cov==4.1.0
requests==2.32.3
